+++ src/command_modules/azure-cli-resource/azure/cli/command_modules/resource/custom.py
@@ -1133,7 +1133,7 @@ def cli_managementgroups_group_create(
         display_name=None,
         parent=None):
     _register_rp(cmd.cli_ctx)
-    parent_id=_get_parent_id_from_parent(parent)
+    parent_id = _get_parent_id_from_parent(parent)
     from azure.mgmt.managementgroups.models import (CreateManagementGroupRequest, CreateManagementGroupDetails, CreateParentGroupInfo)
     create_parent_grp_info = CreateParentGroupInfo(id=parent_id)
     create_mgmt_grp_details = CreateManagementGroupDetails(parent=create_parent_grp_info)
@@ -1144,24 +1144,22 @@ def cli_managementgroups_group_create(
 def cli_managementgroups_group_update_custom_func(
         instance,
         display_name=None,
-        parent=None):
-    instance["display_name"] = display_name
-    instance["parent"] = parent
+        parent_id=None):
+    parent_id = _get_parent_id_from_parent(parent_id)
+    instance.display_name = display_name
+    instance.parent_id = parent_id
     return instance
 
 
 def cli_managementgroups_group_update_get():
-    update_parameters = {'display_name': None, 'parent': None}
+    from azure.mgmt.managementgroups.models import PatchManagementGroupRequest
+    update_parameters = PatchManagementGroupRequest(display_name=None, parent_id=None)
     return update_parameters
 
 
 def cli_managementgroups_group_update_set(
         cmd, client, group_name, parameters=None):
-    _register_rp(cmd.cli_ctx)
-    parent_id=_get_parent_id_from_parent(parameters["parent"])
-    from azure.mgmt.managementgroups.models import PatchManagementGroupRequest
-    patch_mgmt_grp_request = PatchManagementGroupRequest(display_name=parameters["display_name"], parent_id=parent_id)
-    return client.update(group_name, patch_mgmt_grp_request)
+    return client.update(group_name, parameters)
 
 
 def cli_managementgroups_group_delete(cmd, client, group_name):